+++ assets/javascripts/kitten/components/form/color-select/stories.js
@@ -87,7 +87,7 @@ export const Default = ({ value, ...args }) => {
           className="k-u-flex k-u-flex-alignItems-center k-u-flex-justifyContent-end k-u-flex-gap-single"
           style={colorsObject}
         >
-          <Text size="tiny">Rendu&nbsp;:</Text>
+          <Text size="small">Rendu&nbsp;:</Text>
           <Button
             modifier="helium"
             rounded