+++ assets/javascripts/kitten/components/layout/dashboard-layout/stories/with-table.js
@@ -136,7 +136,7 @@ const Row = ({ hasExtraColumn }) => {
 
       <GridTable.Col className="customCol_1">
         <div>
-          <Text size="tiny" weight="regular">
+          <Text size="small" weight="regular">
             <time dateTime="2019-09-12">12/09/2019</time>
           </Text>
           <br />
@@ -171,7 +171,7 @@ const Row = ({ hasExtraColumn }) => {
       </GridTable.Col>
 
       <GridTable.Col className="customCol_3">
-        <Text size="tiny" weight="regular">
+        <Text size="small" weight="regular">
           72&nbsp;€
         </Text>
       </GridTable.Col>
@@ -186,7 +186,7 @@ const Row = ({ hasExtraColumn }) => {
       </GridTable.Col>
 
       <GridTable.Col className="customCol_5">
-        <Text size="tiny" weight="regular">
+        <Text size="small" weight="regular">
           Livraison
         </Text>
       </GridTable.Col>
@@ -206,7 +206,7 @@ const Row = ({ hasExtraColumn }) => {
 
       {hasExtraColumn ? (
         <GridTable.Col className="customCol_7">
-          <Text size="tiny" weight="regular">
+          <Text size="small" weight="regular">
             -
           </Text>
         </GridTable.Col>
@@ -262,13 +262,13 @@ export const StoryWithTable = () => {
             </GridTable.Col>
 
             <GridTable.Col className="customCol_1">
-              <Text weight="regular" color="font1" size="tiny">
+              <Text weight="regular" color="font1" size="small">
                 Date
                 <br className="k-u-hidden@l-up" />
                 <Text
                   weight="regular"
                   color="font1"
-                  size="tiny"
+                  size="small"
                   className="k-u-hidden@l-up"
                 >
                   et Contributeur
@@ -277,38 +277,38 @@ export const StoryWithTable = () => {
             </GridTable.Col>
 
             <GridTable.Col className="customCol_2">
-              <Text weight="regular" color="font1" size="tiny">
+              <Text weight="regular" color="font1" size="small">
                 Contributeur
               </Text>
             </GridTable.Col>
 
             <GridTable.Col className="customCol_3">
-              <Text weight="regular" color="font1" size="tiny">
+              <Text weight="regular" color="font1" size="small">
                 Montant
               </Text>
             </GridTable.Col>
 
             <GridTable.Col className="customCol_4">
-              <Text weight="regular" color="font1" size="tiny">
+              <Text weight="regular" color="font1" size="small">
                 Statut
               </Text>
             </GridTable.Col>
 
             <GridTable.Col className="customCol_5">
-              <Text weight="regular" color="font1" size="tiny">
+              <Text weight="regular" color="font1" size="small">
                 Mode de livraison
               </Text>
             </GridTable.Col>
 
             <GridTable.Col className="customCol_6">
-              <Text weight="regular" color="font1" size="tiny">
+              <Text weight="regular" color="font1" size="small">
                 Statut livraison
               </Text>
             </GridTable.Col>
 
             {hasExtraColumn ? (
               <GridTable.Col className="customCol_7">
-                <Text weight="regular" color="font1" size="tiny">
+                <Text weight="regular" color="font1" size="small">
                   Dernière modification
                 </Text>
               </GridTable.Col>
@@ -337,7 +337,7 @@ export const StoryWithTable = () => {
             className="k-u-flex-alignItems-center"
           >
             <div style={{ flex: '1 0 auto' }} className="k-u-hidden@xs-down">
-              <Text color="background1" size="tiny" weight="regular">
+              <Text color="background1" size="small" weight="regular">
                 Text
               </Text>
             </div>