+++ assets/javascripts/kitten/components/structure/cards/crowdfunding-card/stories.js
@@ -136,12 +136,12 @@ const Info = ({ label, text }) => (
     <Text
       tag="strong"
       weight="bold"
-      size="tiny"
+      size="small"
       className="k-u-inlineFlex@s-down"
     >
       {parseHtml(text)}
     </Text>
-    <Text weight="light" size="tiny">
+    <Text weight="light" size="small">
       {parseHtml(label)}
     </Text>
   </span>