+++ assets/javascripts/kitten/components/typography/horizontal-stroke/index.js
@@ -3,6 +3,7 @@ import PropTypes from 'prop-types'
 import styled, { css } from 'styled-components'
 import COLORS from '../../../constants/colors-config'
 import { pxToRem } from '../../../helpers/utils/typography'
+import { checkDeprecatedSizes } from '../../../helpers/utils/deprecated'
 import classNames from 'classnames'
 import { strokeModifierStylesWithoutMargin } from './common/stroke-modifier-styles'
 
@@ -11,21 +12,25 @@ export const MICRO = css`
   height: ${pxToRem(2)};
 `
 
-export const TINY = css`
+export const SMALL = css`
   width: ${pxToRem(20)};
   height: ${pxToRem(2)};
 `
 
-export const DEFAULT = css`
+export const MEDIUM = css`
   width: ${pxToRem(30)};
   height: ${pxToRem(4)};
 `
 
-export const BIG = css`
+export const LARGE = css`
   width: ${pxToRem(50)};
   height: ${pxToRem(4)};
 `
 
+/* DEPRECATED EXPORTS */
+export const TINY = SMALL
+export const DEFAULT = MEDIUM
+export const BIG = LARGE
 export const HUGE = css`
   width: ${pxToRem(100)};
   height: ${pxToRem(6)};
@@ -42,13 +47,19 @@ export const StyledHorizontalStroke = styled.div`
   &.k-HorizontalStroke--size--tiny {
     ${TINY}
   }
-
-  &.k-HorizontalStroke--size--default {
-    ${DEFAULT}
+  &.k-HorizontalStroke--size--tiny, /* DEPRECATED */
+  &.k-HorizontalStroke--size--small {
+    ${SMALL}
   }
 
-  &.k-HorizontalStroke--size--big {
-    ${BIG}
+  &.k-HorizontalStroke--size--default, /* DEPRECATED */
+  &.k-HorizontalStroke--size--medium {
+    ${MEDIUM}
+  }
+
+  &.k-HorizontalStroke--size--big, /* DEPRECATED */
+  &.k-HorizontalStroke--size--large {
+    ${LARGE}
   }
 
   &.k-HorizontalStroke--size--huge {
@@ -67,6 +78,8 @@ export const HorizontalStroke = ({
   color,
   ...props
 }) => {
+  checkDeprecatedSizes(size)
+
   const modifierClassName = () => {
     if (!modifier) return `k-HorizontalStroke--size--${size}`
 
@@ -96,7 +109,7 @@ export const HorizontalStroke = ({
 }
 
 HorizontalStroke.propTypes = {
-  size: PropTypes.oneOf(['micro', 'tiny', 'default', 'big', 'huge']),
+  size: PropTypes.oneOf(['micro', 'small', 'medium', 'large', 'huge']),
   modifier: PropTypes.oneOf([
     'primary',
     'secondary',
@@ -114,7 +127,7 @@ HorizontalStroke.propTypes = {
 }
 
 HorizontalStroke.defaultProps = {
-  size: 'default',
+  size: 'medium',
   customSize: {
     width: null,
     height: null,